feat: Add "Edit Stock" button and functionality
This commit introduces an "Edit Stock" button to the product editing screen.

Key changes:
- Added a new string resource "btn_edit_stock".
- Integrated an "Edit Stock" button (`FilledTonalIconButton` with an `Edit` icon) in `AddProductComposables.kt` which is visible in edit mode.
- Modified `UpdateProductComposable.kt` to include an `onEditStock` lambda, allowing navigation to the stock details tab when the new button is pressed.
- Updated `UpdateProductUiMappers.kt` to correctly map the product's stock quantity.
- Adjusted `BottomSheetDetailsComposable.kt` to allow setting the initial selected page for the `PagerState`.
- Added a `DetailsOf` enum and modified the `Details` destination in `ProductsDestinations.kt` to support navigating directly to either "Details" or "Stock" tabs.
- Updated `ProductsNavigation.kt` to handle the new `onEditStock` action in `UpdateProductComposableStateImpl`, navigating to the "Stock" tab of the product details. Also, the initial page of the details view is set based on the navigation arguments.
- Changed the `stock` property in `UpdateProductUiModel.kt` from `Float` to `Double`.

+++ demosales-ui/products/src/main/java/com/imecatro/products/ui/update/model/UpdateProductUiModel.kt
@@ -8,7 +8,7 @@ data class UpdateProductUiModel(
     val price: String?,
     val currency: String?,
     val unit: String?,
-    val stock : Float,
+    val stock : Double,
     val imageUri: Uri?,
     val details: String,
     val category: String?
